requirements.txt
streamlit==1.46.0
sqlalchemy==2.0.41
pandas==2.3.0
python-dotenv==1.1.0
openai==1.90.0
pymysql==1.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.0
sqlglot==25.0.0
plotly==5.24.1
streamlit-aggrid==1.0.5
streamlit-ace==0.1.1
cryptography>=41.0.0
google-generativeai>=0.3.0